Styling (WIP)
- Better fonts
- Better colors organization
- Styles on separate file
