+++ stories/index.js
@@ -4,7 +4,7 @@ import { storiesOf } from '@storybook/react';
 import SignUp from '../src/components/001-signup';
 
 storiesOf('DailyUI #001', module).add('SignUp', () => (
-  <div style={{ display: 'flex', justifyContent: 'center' }}>
+  <div style={{ display: 'flex', justifyContent: 'center', padding: '20px' }}>
     <SignUp />
   </div>
 ));
